Add BasePage, BaseClass, logging setup, and log4j2 config
