+++ src/test/java/com/cit/pages/CDCalculatorPageFactory.java
@@ -4,21 +4,12 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
-import org.openqa.selenium.support.PageFactory;
 import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.openqa.selenium.support.ui.WebDriverWait;
 
-import java.time.Duration;
-
-public class CDCalculatorPageFactory {
-
-    private WebDriver driver;
-    private WebDriverWait wait;
+public class CDCalculatorPageFactory extends BasePage{
 
     public CDCalculatorPageFactory(WebDriver driver){
-        this.driver=driver;
-        wait = new WebDriverWait(driver,Duration.ofSeconds(10));
-        PageFactory.initElements(driver,this);
+        super(driver);
     }
 
     @FindBy(xpath = "//input[@id='mat-input-0']")
